requirements.txt
fastapi[all]
line-bot-sdk
spotipy
gunicorn
requests
python-dotenv